+++ lib/ckb/utils.rb
@@ -48,5 +48,19 @@ module CKB
 
       raise "Can't convert to hex string!"
     end
+
+    def self.sudt_amount!(output_data)
+      return 0 if output_data == "0x"
+      output_data_bin = CKB::Utils.hex_to_bin(output_data)
+      raise "Invalid sUDT amount" if output_data_bin.bytesize < 16
+
+      values = output_data_bin[0..15].unpack("Q<Q<")
+      (values[1] << 64) | values[0]
+    end
+
+    def self.generate_sudt_amount(sudt_amount)
+      values = [sudt_amount & 0xFFFFFFFFFFFFFFFF, (sudt_amount >> 64) & 0xFFFFFFFFFFFFFFFF]
+      CKB::Utils.bin_to_hex(values.pack("Q<Q<"))
+    end
   end
 end