+++ lib/ckb/api.rb
@@ -243,6 +243,7 @@ module CKB
 
     # @param lock_hash [String]
     def deindex_lock_hash(lock_hash)
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       rpc.deindex_lock_hash(lock_hash)
     end
 
@@ -253,12 +254,14 @@ module CKB
     #
     # @return [Types::LiveCell[]]
     def get_live_cells_by_lock_hash(lock_hash, page, per, reverse_order: false)
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       result = rpc.get_live_cells_by_lock_hash(lock_hash, page, per, reverse_order)
       result.map { |live_cell| Types::LiveCell.from_h(live_cell) }
     end
 
     # @return [Types::LockHashIndexState[]]
     def get_lock_hash_index_states
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       result = rpc.get_lock_hash_index_states
       result.map { |state| Types::LockHashIndexState.from_h(state) }
     end
@@ -270,6 +273,7 @@ module CKB
     #
     # @return [Types::CellTransaction[]]
     def get_transactions_by_lock_hash(lock_hash, page, per, reverse_order: false)
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       result = rpc.get_transactions_by_lock_hash(lock_hash, page, per, reverse_order)
       result.map { |cell_tx| Types::CellTransaction.from_h(cell_tx) }
     end
@@ -279,6 +283,7 @@ module CKB
     #
     # @return [Types::LockHashIndexState]
     def index_lock_hash(lock_hash, index_from: 0)
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       state = rpc.index_lock_hash(lock_hash, index_from)
       Types::LockHashIndexState.from_h(state)
     end
@@ -287,6 +292,7 @@ module CKB
     #
     # @return [CKB::Types::LockHashCapacity]
     def get_capacity_by_lock_hash(lock_hash)
+      warn "[DEPRECATION] This method is deprecated since version 0.36.0 for reasons of flexibility, please use ckb-indexer as an alternate solution"
       result = rpc.get_capacity_by_lock_hash(lock_hash)
       Types::LockHashCapacity.from_h(result) unless result.nil?
     end